Refactor duplicate code across translation scripts into translation_utils.py
